+++ lib/secretary/has_secretary.rb
@@ -3,15 +3,25 @@ module Secretary
     extend ActiveSupport::Concern
 
     module ClassMethods
-      # Check if a class is versioned
+      # (Boolean) Check if a class is versioned
+      # Story.has_secretary? # => true or false
       def has_secretary?
         !!@_has_secretary
       end
 
-      # Apply to any class that should be versioned
-      def has_secretary
+      # Declare that this class shoudl be versioned.
+      # 
+      # Options:
+      # * `on` : Array of Strings which specifies which attributes should be
+      #   versioned.
+      # * `except` : Array of Strings which specifies which attributes should
+      #   NOT be versioned.
+      def has_secretary(options={})
         @_has_secretary = true
         Secretary.versioned_models.push self.name
+
+        self.versioned_attributes   = options[:on]     if options[:on]
+        self.unversioned_attributes = options[:except] if options[:except]
 
         has_many :versions,
           :class_name   => "Secretary::Version",