+++ spec/internal/app/models/person.rb
@@ -1,14 +1,10 @@
 class Person < ActiveRecord::Base
-  has_secretary
+  has_secretary except: ["name", "ethnicity"]
 
   belongs_to :location
-
   has_many :animals
-  accepts_nested_attributes_for :animals, allow_destroy: true
-
   has_many :hobbies
 
+  accepts_nested_attributes_for :animals, allow_destroy: true
   tracks_association :animals, :hobbies
-
-  self.unversioned_attributes = ["name", "ethnicity"]
 end