+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+## 1.0.0.beta2
+#### Additions
+* Added better way to declare attribute inclusion/exclusion:
+  `has_secretary on: []` and `has_secretary except: []`. This ensures that
+  attributes will be declared before any `tracks_association` are called.
 ## 1.0.0.beta1 (2013-11-05)
 * Beta for stable 1.0 release.
+
 
 ## 0.0.1
 * Initial release